+++ R/testdata.R
@@ -14,7 +14,7 @@
 #' data(coloc_test_data)
 #' names(coloc_test_data)
 #' str(coloc_test_data$D1)
-#' check.dataset(coloc_test_data$D1) # should return NULL if data structure is ok
+#' check_dataset(coloc_test_data$D1) # should return NULL if data structure is ok
 "coloc_test_data"
 
 if(FALSE) {